+++ src/Providers/Editor/EditorControllerProvider.cs
@@ -1,7 +1,4 @@
-﻿using System.IO;
-using System.Reflection;
-
-namespace DialToolsForVS
+﻿namespace DialToolsForVS
 {
     [DialControllerProvider(4)]
     internal class EditorControllerProvider : IDialControllerProvider
@@ -10,9 +7,7 @@ namespace DialToolsForVS
 
         public IDialController TryCreateController(IDialControllerHost host)
         {
-            string folder = Path.GetDirectoryName(Assembly.GetExecutingAssembly().Location);
-            string iconFilePath = Path.Combine(folder, "Providers\\Editor\\icon.png");
-
+            string iconFilePath = VsHelpers.GetFileInVsix("Providers\\Editor\\icon.png");
             host.AddMenuItem(Moniker, iconFilePath);
 
             return new EditorController();