+++ src/Providers/Scroll/ScrollControllerProvider.cs
@@ -1,7 +1,4 @@
-﻿using System.IO;
-using System.Reflection;
-
-namespace DialToolsForVS
+﻿namespace DialToolsForVS
 {
     [DialControllerProvider(order: 1)]
     internal class ScrollControllerProvider : IDialControllerProvider
@@ -9,9 +6,7 @@ namespace DialToolsForVS
         public const string Moniker = "Scroll";
         public IDialController TryCreateController(IDialControllerHost host)
         {
-            string folder = Path.GetDirectoryName(Assembly.GetExecutingAssembly().Location);
-            string iconFilePath = Path.Combine(folder, "Providers\\Scroll\\icon.png");
-
+            string iconFilePath = VsHelpers.GetFileInVsix("Providers\\Scroll\\icon.png");
             host.AddMenuItem(Moniker, iconFilePath);
 
             return new ScrollController();